Working on transfering changes to more of an OOP standard
