+++ .vscode/settings.json
@@ -11,6 +11,7 @@
     "pyaudio",
     "pygame",
     "rects",
+    "Renderable",
     "Shrikhand",
     "VIDEORESIZE"
   ]
