+++ src/dev.py
@@ -1,0 +1,2 @@
+def todo():
+    raise NotImplementedError("This method is not implemented yet.")
