+++ src/qwerty_to_midi.csv
@@ -1,0 +1,19 @@
+a,60
+w,61
+s,62
+e,63
+d,64
+f,65
+t,66
+g,67
+y,68
+h,69
+u,70
+j,71
+k,72
+o,73
+l,74
+p,75
+;,76
+',77
+],78
