+++ src_old/note_visuals.py
